Add files via upload
One of the main problems with GDP per capita is that it doesn't account for any inequality within a society. GDP counts "bads" as well as "goods." When an earthquake hits and requires rebuilding, GDP increases. GDP makes no adjustment for leisure time. GDP only counts goods that pass through official, organized markets, so it misses home production and black market activity.
The implementation of this can be done by using Linear Regression. Linear Regression is a machine learning algorithm based on supervised learning. Itperforms a regression task. Regression models a target prediction value based on independent variables. It is mostly used for finding out the relationship between variables and forecasting.We are going to predict the Gross domestic product for the future years by
using the technologies like pandas, matplotlib, Scikit-learn and The Jupyter